requirements.txt
SQLAlchemy==1.4.37
asyncio==3.4.3
bcrypt==3.2.2
fastapi==0.61.2
fastapi-cache==0.1.0
fastapi-mail==0.4.1
uvicorn==0.13.4
websockets==8.1
redis==4.3.1
python-dotenv==0.20.0
PyJWT==2.4.0
psycopg2-binary==2.9.3
